+++ workers/unity/Packages/io.improbable.gdk.core/.codegen/Source/Generators/Parts/MetaclassGeneratorPart.cs
@@ -1,0 +1,29 @@
+using System.Collections.Generic;
+using Improbable.Gdk.CodeGeneration.Model.Details;
+
+namespace Improbable.Gdk.CodeGenerator
+{
+    public partial class MetaclassGenerator
+    {
+        private UnityComponentDetails details;
+        private string qualifiedNamespace;
+
+        public string Generate(UnityComponentDetails details, string package)
+        {
+            qualifiedNamespace = package;
+            this.details = details;
+
+            return TransformText();
+        }
+
+        private UnityComponentDetails GetComponentDetails()
+        {
+            return details;
+        }
+
+        private IReadOnlyList<UnityCommandDetails> GetCommandDetailsList()
+        {
+            return details.CommandDetails;
+        }
+    }
+}