+++ workers/unity/Packages/io.improbable.gdk.core/Components/IComponentMetaclass.cs
@@ -1,0 +1,26 @@
+using System;
+using Improbable.Gdk.Core.Commands;
+
+namespace Improbable.Gdk.Core
+{
+    public interface IComponentMetaclass
+    {
+        uint ComponentId { get; }
+        string Name { get; }
+
+        Type Data { get; }
+        Type Snapshot { get; }
+        Type Update { get; }
+
+        Type ReplicationHandler { get; }
+        Type Serializer { get; }
+        Type DiffDeserializer { get; }
+
+        Type DiffStorage { get; }
+        Type ViewStorage { get; }
+        Type EcsViewManager { get; }
+        Type DynamicInvokable { get; }
+
+        ICommandMetaclass[] Commands { get; }
+    }
+}